document_intelligence/requirements.txt
boto3==1.34.0
python-dotenv==1.0.0
gradio==4.44.0
PyPDF2==3.0.1
pdfplumber==0.10.3
Pillow==10.1.0
pytesseract==0.3.10
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4
requests==2.31.0
crewai==0.28.0
pydantic==2.5.0
